+++ packages/db/drizzle/0199_small_kat_farrell.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "autonomy_contracts" ADD COLUMN "capabilities" text[] DEFAULT '{}'::text[] NOT NULL;